fix: add handling for two-elev zero gate ratings
Within the CWMS database, gate ratings for the "zero" settings are often
entered with only two elevations: the minimum and maximum, both set to
zero flow. These condensed ratings were causing errors within the
get_interp_list function. The function has been updated to identify and
expand these ratings to contain a number of zero values equal to the
number of flow values present in other settings for the gate.