Corrected CaveRun failure to load
Updated Tailwater json entries to allow for different gage sites for stage
and temperature due Cave Run using farmers for stage and farmers at hwy60 for temperature

Changed the submission error check order and added a try except to catch
type exceptions.

Resolved the failure at submission that caused the program to freeze (it
was the result of referencing a varible with out the self. before)

Fixed the reference lines for plots

Started the process of making the exception handling more specific to
help with debugging... work in progress

Note this version is still in test mode.